Adder folders in ZIP backup. Small fixes

+++ app/src/main/java/com/privategallery/akscorp/privategalleryandroid/Essentials/Image.kt
@@ -10,5 +10,5 @@ import java.io.Serializable
 
 data class Image(var id: Long? = null, var name: String? = null, var localPath:
 String? = null, var albumId: Long? = null, var extension: String? = null,
-                 var addedTime: Long? = null, var height: Long? = null, var width: Long? = null)
+                 var addedTime: Long? = null, var height: Int? = null, var width: Int? = null)
     : Serializable
